Create Main.py
The definition of image encryption used here is an extension of data encryption in general: using the bitwise XOR operation of the original image pixels and the random key image pixels, with the key image being either a pseudo-random stream cipher or the quantum random stream cipher or anti-correlated entangled information shared over a secure channel. In either case we Perform bitwise XOR operation on the encrypted image and the key image. It can be seen from the image encryption and decryption that they are all the same operation. According to the above bitwise XOR operation, we assume: xor(a, b) = c You can get: xor(c, b) = a Or: xor(c, a) = b In summary, we assume that a is the original image data and b is the key, then c calculated by xor(a, c) is the encrypted ciphertext. This is a simple summary of the encryption and decryption process as used in the coding in this repository. Encryption process: Perform a bitwise XOR operation on the image a and the key b to complete the encryption and obtain the ciphertext c. Decryption process: Perform a bitwise XOR operation on the ciphertext c and the key b, complete the decryption, and get the image a. We can use our quantum random numbers generator in 2 ways to create our image encryption key: (1) as a random number generator seed source (2) using the random superposition of the H and V modes We can also use the shared set of correlated images, captured using the single CCD, from our entangled photon source with Alice getting one half and Bob getting the anti-correlated half. This provides the perfect key, with the quantum images shared over a separate channel hidden from the encrypted classical images. The file exchange channel is 2 way: Alice can use Her key to encrypt the image, Bob can use His key to decrypt the image OR Bob can use His key to encrypt the image, Alice can use Her key to decrypt the image. The XOR Cipher in this use can also be extended as a component in more complex overlay network ciphers if need be however for computational efficiency it is not necessary. It is just as effectual to have 1 quantum cipher as many, so in effect the system is completely hidden, by virtue of hidden variables, and is encrypted in an information condensate.